Restructured KeyFactors & Comments module
